完善分类列表
郑肃

+++ your-project-name/database/migrations/2018_03_13_174548_create_goodtypes_table.php
@@ -15,9 +15,9 @@ class CreateGoodtypesTable extends Migration
         if (!Schema::hasTable('goodtypes')) {
             Schema::create('goodtypes', function (Blueprint $table) {
                 $table->increments('id');
-                $table->tinyInteger('one_id');
-                $table->tinyInteger('two_id');
-                $table->tinyInteger('nav_id');
+                $table->int('one_id');
+                $table->int('two_id');
+                $table->int('nav_id');
                 $table->string('gt_name',255);//商品属性名
                 $table->timestamps();
             });
