+++ your-project-name/database/migrations/2018_03_19_210706_create_goodtypevals_table.php
@@ -15,10 +15,10 @@ class CreateGoodtypevalsTable extends Migration
         if (!Schema::hasTable('goodtypevals')) {
             Schema::create('goodtypevals', function (Blueprint $table) {
                 $table->increments('id');
-                $table->tinyInteger('yiji_id');
-                $table->tinyInteger('erji_id');
-                $table->tinyInteger('sanji_id');
-                $table->tinyInteger('gt_id');
+                $table->int('yiji_id');
+                $table->int('erji_id');
+                $table->int('sanji_id');
+                $table->int('gt_id');
                 $table->string('gtv_name',255);//商品属性名
                 $table->timestamps();
             });
